+++ templates/antimicrobial_resistance.sh
@@ -8,7 +8,7 @@ mkdir -p ${LOG_DIR}
 # Print captured STDERR incase of exit
 function print_stderr {
     cat .command.err 1>&2
-    ls ${LOG_DIR}/ | grep ".err" | xargs -I {} cat {} 1>&2
+    ls ${LOG_DIR}/ | grep ".err" | xargs -I {} cat ${LOG_DIR}/{} 1>&2
 }
 trap print_stderr EXIT