+++ templates/call_variants_auto.sh
@@ -11,7 +11,7 @@ snippy --version >> ${LOG_DIR}/!{task.process}.versions 2>&1
 # Print captured STDERR incase of exit
 function print_stderr {
     cat .command.err 1>&2
-    ls ${LOG_DIR}/ | grep ".err" | xargs -I {} cat {} 1>&2
+    ls ${LOG_DIR}/ | grep ".err" | xargs -I {} cat ${LOG_DIR}/{} 1>&2
 }
 trap print_stderr EXIT